pyproject.toml
[project]
name = "hacktheback"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.13.3",
    "fastapi[standard]>=0.115.2",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.9",
    "pyjwt>=2.9.0",
    "sqlmodel>=0.0.22",
]
